nvmf: Add a balanced qpair scheduling by poll groups
The current connection scheduling mechanism (RoundRobin) doesn't take into account a qpair type and assigns each new qpair to the next poll group. As a side effect there might occur a disbalance when some poll group handles more IO qpairs than others. The patch adds a configuration parameter "ConnSched" for Nvmf section with the following possible values: "HostIp", "RoundRobin", "TransportSpec" (the new type of the scheduling). If "TransportSpec" type has been configured then a poll group selection logic will be passed to the transport subsystem.
In RDMA transport it is possible to get the qpair type before the controller creation using a private data from the rdma_cm event, this allows to schedule admin and IO qpairs in the balanced way. In TCP transport it is impossible to get the qpair type, so the original "RoundRobin" logic is used.

Signed-off-by: Alexey Marchuk <alexeymar@mellanox.com>
Signed-off-by: Sasha Kotchubievsky <sashakot@mellanox.com>
Signed-off-by: Evgenii Kochetov <evgeniik@mellanox.com>
Change-Id: I1d55f69c14c810d3da831cca17ebd62e6ef200c7